apps/backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-mongodb-backend==5.2.1
django-cors-headers==4.9.0
python-dotenv==1.1.1
bleach==6.2.0
